@@ -20,6 +20,9 @@ android {
             buildConfigField 'String', 'HEALTHY_ESTHER_DROPBOX_KEY', "\"${dropboxKey}\""
             resValue 'string', 'HEALTHY_ESTHER_DROPBOX_KEY', "db-${dropboxKey}"
             buildConfigField 'String', 'HEALTHY_ESTHER_DROPBOX_SECRET', "\"${dropboxSecret}\""
+            minifyEnabled true
+            proguardFiles getDefaultProguardFile('proguard-android.txt'),
+                    'proguard-rules.pro'
         }
     }
     signingConfigs {
@@ -30,26 +33,33 @@ android {
             storePassword System.getenv('HEALTHY_ESTHER_STORE_PASSWORD')
         }
     }
-    compileSdkVersion 25
-    buildToolsVersion '25.0.0'
+    testOptions {
+        unitTests {
+            includeAndroidResources true
+        }
+    }
+    compileSdkVersion 27
+    buildToolsVersion '27.0.3'
     defaultConfig {
         applicationId "com.robwilliamson.healthyesther"
-        jackOptions {
-            enabled true
-        }
-        minSdkVersion 19
-        targetSdkVersion 25
-        versionCode 28
-        versionName "0.10.0"
+        minSdkVersion 26
+        targetSdkVersion 27
+        versionCode 29
+        versionName "0.11.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         //noinspection GroovyAssignabilityCheck
         signingConfig signingConfigs.release
+        multiDexEnabled true
     }
     packagingOptions {
-        exclude 'LICENSE.txt'
+        exclude 'META-INF/DEPENDENCIES'
         exclude 'META-INF/LICENSE'
         exclude 'META-INF/LICENSE.txt'
+        exclude 'META-INF/license.txt'
         exclude 'META-INF/NOTICE'
+        exclude 'META-INF/NOTICE.txt'
+        exclude 'META-INF/notice.txt'
+        exclude 'META-INF/ASL2.0'
     }
     dexOptions {
     }
@@ -75,20 +85,21 @@ task checkEnvironnment {
 tasks.preBuild.dependsOn(":dbcodemodel:sqlite2Java")
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile group: 'com.dropbox.core', name: 'dropbox-core-sdk', version: '2.0-beta-7'
-    compile 'com.squareup.okhttp:okhttp:2.4.0'
-    compile 'com.android.support:support-v4:25.2.0'
-    compile 'com.android.support:appcompat-v7:25.2.0'
-    compile 'net.danlew:android.joda:2.7.1'
-    compile 'com.jjoe64:graphview:3.1.3'
-    compile 'org.apache.commons:commons-math3:3.6.1'
-    androidTestCompile 'com.android.support.test.espresso:espresso-core:2.2.2'
-    androidTestCompile 'com.android.support.test:runner:0.5'
-    androidTestCompile 'com.android.support:support-annotations:25.2.0'
-    testCompile 'org.robolectric:robolectric:3.0'
-    testCompile 'org.mockito:mockito-core:2.0.31-beta'
-    testCompile 'org.hamcrest:hamcrest-all:1.3'
-    compile project(':dbcodemodel')
-    compile project(':includes')
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'com.google.code.findbugs:jsr305:2.0.1'
+    implementation 'com.dropbox.core:dropbox-core-sdk:3.0.0'
+    implementation 'com.android.support:support-v4:27.1.0'
+    implementation 'com.android.support:appcompat-v7:27.1.0'
+    implementation 'com.jjoe64:graphview:3.1.3'
+    implementation 'org.apache.commons:commons-math3:3.6.1'
+    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    androidTestImplementation 'com.android.support.test:runner:1.0.1'
+    androidTestImplementation 'com.android.support:support-annotations:27.1.0'
+    implementation 'junit:junit:4.12'
+    testImplementation 'org.robolectric:robolectric:3.8'
+    testImplementation 'org.robolectric:shadows-support-v4:3.0'
+    testImplementation 'org.mockito:mockito-core:2.13.0'
+    testImplementation 'org.hamcrest:hamcrest-all:1.3'
+    implementation project(':dbcodemodel')
+    implementation project(':includes')
 }
