@@ -1,6 +1,6 @@
-/**
-  * © Robert Williamson 2014-2016.
-  * This program is distributed under the terms of the GNU General Public License.
+/*
+   © Robert Williamson 2014-2016.
+   This program is distributed under the terms of the GNU General Public License.
   */
 package com.robwilliamson.healthyesther.edit;
 
@@ -12,9 +12,7 @@ import com.robwilliamson.healthyesther.db.generated.EventTable;
 import com.robwilliamson.healthyesther.db.generated.HealthDatabase;
 import com.robwilliamson.healthyesther.db.generated.NoteEventTable;
 import com.robwilliamson.healthyesther.db.generated.NoteTable;
-import com.robwilliamson.healthyesther.db.includes.Database;
 import com.robwilliamson.healthyesther.db.includes.DateTime;
-import com.robwilliamson.healthyesther.db.includes.Transaction;
 import com.robwilliamson.healthyesther.db.includes.TransactionExecutor;
 import com.robwilliamson.healthyesther.db.includes.WhereContains;
 import com.robwilliamson.healthyesther.db.integration.DatabaseAccessor;
@@ -57,7 +55,7 @@ public class NoteEventActivity extends AbstractEditActivity
 
         note.setAlwaysCreate(true);
 
-        list.add(new Pair<>((EditFragment) note, NOTE_TAG));
+        list.add(new Pair<>(note, NOTE_TAG));
         list.add(new Pair<>(event, EVENT_TAG));
 
         return list;
@@ -65,34 +63,26 @@ public class NoteEventActivity extends AbstractEditActivity
 
     @Override
     protected TransactionExecutor.Operation onModifySelected() {
-        return new TransactionExecutor.Operation() {
-            @Override
-            public void doTransactionally(@Nonnull Database database, @Nonnull Transaction transaction) {
-                EventTable.Row row = Utils.checkNotNull(getEventFragment().getRow());
-                row.setTypeId(EventTypeTable.NOTE.getId());
-                row.applyTo(transaction);
+        return (database, transaction) -> {
+            EventTable.Row row = Utils.checkNotNull(getEventFragment().getRow());
+            row.setTypeId(EventTypeTable.NOTE.getId());
+            row.applyTo(transaction);
 
-                NoteTable.Row note = Utils.checkNotNull(getNoteFragment().getRow());
-                note.applyTo(transaction);
+            NoteTable.Row note = Utils.checkNotNull(getNoteFragment().getRow());
+            note.applyTo(transaction);
 
-                NoteEventTable.Row noteEventRow = DatabaseAccessor.NOTE_EVENT_TABLE.select0Or1(
-                        database, WhereContains.and(
-                                WhereContains.foreignKey(NoteEventTable.EVENT_ID, row.getNextPrimaryKey().getId()),
-                                WhereContains.foreignKey(NoteEventTable.NOTE_ID, note.getNextPrimaryKey().getId())));
+            NoteEventTable.Row noteEventRow = DatabaseAccessor.NOTE_EVENT_TABLE.select0Or1(
+                    database, WhereContains.and(
+                            WhereContains.foreignKey(NoteEventTable.EVENT_ID, row.getNextPrimaryKey().getId()),
+                            WhereContains.foreignKey(NoteEventTable.NOTE_ID, note.getNextPrimaryKey().getId())));
 
-                if (noteEventRow == null) {
-                    noteEventRow = new NoteEventTable.Row(row.getNextPrimaryKey(), note.getNextPrimaryKey());
-                }
-
-                noteEventRow.applyTo(transaction);
-
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        finish();
-                    }
-                });
+            if (noteEventRow == null) {
+                noteEventRow = new NoteEventTable.Row(row.getNextPrimaryKey(), note.getNextPrimaryKey());
             }
+
+            noteEventRow.applyTo(transaction);
+
+            runOnUiThread(this::finish);
         };
     }
 
@@ -105,24 +95,16 @@ public class NoteEventActivity extends AbstractEditActivity
 
         getEventFragment().setRow(event);
 
-        getExecutor().perform(new TransactionExecutor.Operation() {
-            @Override
-            public void doTransactionally(@Nonnull Database database, @Nonnull Transaction transaction) {
-                NoteEventTable.Row noteEvent = HealthDatabase.NOTE_EVENT_TABLE.select1(
-                        database,
-                        foreignKey(NoteEventTable.EVENT_ID, event.getConcretePrimaryKey().getId()));
+        getExecutor().perform((database, transaction) -> {
+            NoteEventTable.Row noteEvent = HealthDatabase.NOTE_EVENT_TABLE.select1(
+                    database,
+                    foreignKey(NoteEventTable.EVENT_ID, event.getConcretePrimaryKey().getId()));
 
-                final NoteTable.Row note = HealthDatabase.NOTE_TABLE.select1(
-                        database,
-                        noteEvent.getConcretePrimaryKey().getNoteId());
+            final NoteTable.Row note = HealthDatabase.NOTE_TABLE.select1(
+                    database,
+                    noteEvent.getConcretePrimaryKey().getNoteId());
 
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        getNoteFragment().setRow(note);
-                    }
-                });
-            }
+            runOnUiThread(() -> getNoteFragment().setRow(note));
         });
     }
 
