@@ -4,10 +4,10 @@
   */
 package com.robwilliamson.healthyesther.util.time;
 
-import org.joda.time.DateTime;
-import org.joda.time.Duration;
-import org.joda.time.ReadableInstant;
-
+import java.time.Duration;
+import java.time.Instant;
+import java.time.ZonedDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
@@ -20,8 +20,8 @@ public class RangeSet extends TimeRegion {
         mTimeRegions = new HashSet<>(Arrays.asList(timeRegions));
     }
 
-    private static DateTime getFrom(TimeRegion... timeRegions) {
-        DateTime from = null;
+    private static ZonedDateTime getFrom(TimeRegion... timeRegions) {
+        ZonedDateTime from = null;
 
         for (TimeRegion region : timeRegions) {
             if (from == null || region.from.isBefore(from)) {
@@ -32,8 +32,8 @@ public class RangeSet extends TimeRegion {
         return from;
     }
 
-    private static DateTime getTo(TimeRegion... timeRegions) {
-        DateTime to = null;
+    private static ZonedDateTime getTo(TimeRegion... timeRegions) {
+        ZonedDateTime to = null;
 
         for (TimeRegion region : timeRegions) {
             if (to == null || region.to.isAfter(to)) {
@@ -44,8 +44,8 @@ public class RangeSet extends TimeRegion {
         return to;
     }
 
-    private static boolean afterOrEqualTo(ReadableInstant lhs, ReadableInstant rhs) {
-        return lhs.isAfter(rhs) || lhs.isEqual(rhs);
+    private static boolean afterOrEqualTo(Instant lhs, Instant rhs) {
+        return lhs.isAfter(rhs) || lhs.equals(rhs);
     }
 
     @Override
@@ -71,7 +71,7 @@ public class RangeSet extends TimeRegion {
     }
 
     @Override
-    public boolean contains(ReadableInstant instant, Comparison comparison) {
+    public boolean contains(Instant instant, Comparison comparison) {
         for (TimeRegion subRegion : mTimeRegions) {
             if (subRegion.contains(instant, comparison)) {
                 return true;
@@ -82,28 +82,32 @@ public class RangeSet extends TimeRegion {
 
     @Override
     public RangeSet startingFrom(int year, int monthOfYear, int dayOfMonth) {
-        Duration shift = Duration.millis(from.withDate(year, monthOfYear, dayOfMonth).getMillis() - from.getMillis());
+        Duration shift = Duration.of(
+                from.withYear(year)
+                        .withMonth(monthOfYear)
+                        .withDayOfMonth(dayOfMonth).toInstant().toEpochMilli()
+                        - from.toInstant().toEpochMilli(), ChronoUnit.MILLIS);
         Set<TimeRegion> regions = new HashSet<>(mTimeRegions.size());
 
         for (TimeRegion region : mTimeRegions) {
-            DateTime newFrom = region.from.plus(shift);
-            regions.add(region.startingFrom(newFrom.getYear(), newFrom.getMonthOfYear(), newFrom.getDayOfMonth()));
+            ZonedDateTime newFrom = region.from.plus(shift);
+            regions.add(region.startingFrom(newFrom.getYear(), newFrom.getMonthValue(), newFrom.getDayOfMonth()));
         }
 
         return new RangeSet(regions.toArray(new TimeRegion[]{}));
     }
 
-    public DateTime getEdgeAfter(DateTime instant) {
-        DateTime edge = null;
+    public ZonedDateTime getEdgeAfter(Instant instant) {
+        ZonedDateTime edge = null;
 
         for (TimeRegion subRegion : mTimeRegions) {
-            DateTime nextEdge = null;
+            ZonedDateTime nextEdge = null;
             if (subRegion instanceof RangeSet) {
                 nextEdge = ((RangeSet) subRegion).getEdgeAfter(instant);
             } else {
-                if (afterOrEqualTo(subRegion.from, instant)) {
+                if (afterOrEqualTo(subRegion.from.toInstant(), instant)) {
                     nextEdge = subRegion.from;
-                } else if (afterOrEqualTo(subRegion.to, instant)) {
+                } else if (afterOrEqualTo(subRegion.to.toInstant(), instant)) {
                     nextEdge = subRegion.to;
                 }
             }
