@@ -1,6 +1,6 @@
-/**
-  * © Robert Williamson 2014-2016.
-  * This program is distributed under the terms of the GNU General Public License.
+/*
+   © Robert Williamson 2014-2016.
+   This program is distributed under the terms of the GNU General Public License.
   */
 package com.robwilliamson.healthyesther.unit.com.robwilliamson.healthyesther.db.utils;
 
@@ -10,33 +10,49 @@ import android.support.test.runner.AndroidJUnit4;
 
 import com.robwilliamson.healthyesther.db.Utils;
 
-import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.TimeZone;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(AndroidJUnit4.class)
 public class TimeTest {
-    private final DateTime UTC = new DateTime(2015, 3, 22, 8, 56, 25, 0, DateTimeZone.UTC);
+    private final ZonedDateTime UTC = ZonedDateTime.of(
+            2015,
+            3,
+            22,
+            8,
+            56,
+            25,
+            0,
+            ZoneOffset.UTC);
     private final String UTC_STRING = "2015-03-22T08:56:25 +00:00";
-    private final DateTime CET = new DateTime(2015, 3, 22, 8, 56, 25, 0, DateTimeZone.forID("+0100"));
+    private final ZonedDateTime CET = ZonedDateTime.of(
+            2015,
+            3,
+            22,
+            8,
+            56,
+            25,
+            0,
+            ZoneId.of("+0100"));
     private final String CET_STRING = "2015-03-22T08:56:25 +01:00";
 
     @Test
     public void testBundling() {
         Bundle bundle = new Bundle();
         String name = "testTime";
-        DateTime input = DateTime.now().withZone(
-                DateTimeZone.forOffsetMillis(
-                        TimeZone.getDefault().getRawOffset()));
-        System.out.println(input.getZone());
+        ZonedDateTime input = ZonedDateTime.now()
+                .withZoneSameLocal(ZoneId.systemDefault())
+                .withFixedOffsetZone()
+                .withNano(0);
         Utils.Time.bundle(bundle, name, input);
-        DateTime output = Utils.Time.unBundle(bundle, name);
+        ZonedDateTime output = Utils.Time.unBundle(bundle, name);
         assertTrue(input.isEqual(output));
         assertEquals(input.getZone(), output.getZone());
         assertTrue(input.getZone().equals(output.getZone()));
@@ -44,9 +60,8 @@ public class TimeTest {
 
     @Test
     public void testLocalNow() {
-        DateTime localNow = Utils.Time.localNow();
-        assertEquals(localNow.getZone(), DateTimeZone.forTimeZone(
-                TimeZone.getDefault()));
+        ZonedDateTime localNow = Utils.Time.localNow();
+        assertEquals(localNow.getZone(), ZoneId.systemDefault());
     }
 
     @Test
@@ -79,8 +94,8 @@ public class TimeTest {
         fromDatabaseString(CET_STRING, CET);
     }
 
-    private void fromDatabaseString(String string, DateTime expected) {
-        DateTime actual = Utils.Time.fromDatabaseString(string);
+    private void fromDatabaseString(String string, ZonedDateTime expected) {
+        ZonedDateTime actual = Utils.Time.fromDatabaseString(string);
         assertEquals(expected.getZone(), actual.getZone());
         assertTrue("Expect " + expected + " to equal " + actual, actual.isEqual(expected));
     }
@@ -95,8 +110,8 @@ public class TimeTest {
         checkFromLocalString(CET_STRING, CET);
     }
 
-    private void checkFromLocalString(String string, DateTime expected) {
-        DateTime actual = Utils.Time.fromLocalString(string);
+    private void checkFromLocalString(String string, ZonedDateTime expected) {
+        ZonedDateTime actual = Utils.Time.fromLocalString(string);
         assertEquals(expected.getZone(), actual.getZone());
         assertTrue("Expect " + expected + " to equal " + actual, actual.isEqual(expected));
     }
@@ -104,6 +119,6 @@ public class TimeTest {
     @Test
     public void testToBootRealTimeElapsedMillis() {
         long millisTime = SystemClock.elapsedRealtime();
-        assertEquals(millisTime, Utils.Time.toBootRealTimeElapsedMillis(DateTime.now()), 1000);
+        assertEquals(millisTime, Utils.Time.toBootRealTimeElapsedMillis(ZonedDateTime.now()), 1000);
     }
 }
