@@ -1,6 +1,6 @@
-/**
-  * © Robert Williamson 2014-2016.
-  * This program is distributed under the terms of the GNU General Public License.
+/*
+   © Robert Williamson 2014-2016.
+   This program is distributed under the terms of the GNU General Public License.
   */
 package com.robwilliamson.healthyesther.unit.com.robwilliamson.healthyesther.reminder;
 
@@ -11,15 +11,16 @@ import android.test.mock.MockContext;
 import com.robwilliamson.healthyesther.reminder.TimingModel;
 import com.robwilliamson.healthyesther.util.time.Range;
 
-import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
-import org.joda.time.Duration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.time.Duration;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
 
 import static com.robwilliamson.healthyesther.unit.Assert.assertIsEqual;
 import static org.junit.Assert.assertEquals;
@@ -27,20 +28,61 @@ import static org.junit.Assert.assertNull;
 
 @RunWith(AndroidJUnit4.class)
 public class TimingModelTest {
-    private static final DateTime MORNING_8AM_21 = new DateTime(2015, 3, 21, 8, 0).withZone(DateTimeZone.UTC);
-    private static final DateTime MORNING_21 = new DateTime(2015, 3, 21, 7, 0);
-    private static final DateTime MIDDAY_21 = new DateTime(2015, 3, 21, 12, 0);
-    private static final DateTime EVENING_21 = new DateTime(2015, 3, 21, 22, 0);
-    private static final DateTime MIDNIGHT_21 = new DateTime(2015, 3, 21, 0, 0);
+    private static final ZonedDateTime MORNING_8AM_21 = ZonedDateTime.of(
+            2015,
+            3,
+            21,
+            8,
+            0,
+            0,
+            0,
+            ZoneOffset.UTC);
+    private static final ZonedDateTime MORNING_21 = ZonedDateTime.of(
+            2015,
+            3,
+            21,
+            7,
+            0,
+            0,
+            0,
+            ZoneId.systemDefault());
+    private static final ZonedDateTime MIDDAY_21 = ZonedDateTime.of(
+            2015,
+            3,
+            21,
+            12,
+            0,
+            0,
+            0,
+            ZoneId.systemDefault());
+    private static final ZonedDateTime EVENING_21 = ZonedDateTime.of(
+            2015,
+            3,
+            21,
+            22,
+            0,
+            0,
+            0,
+            ZoneId.systemDefault());
+    private static final ZonedDateTime MIDNIGHT_21 = ZonedDateTime.of(
+            2015,
+            3,
+            21,
+            0,
+            0,
+            0,
+            0,
+            ZoneId.systemDefault());
     private static final Range ALLOWED = new Range(MORNING_21, EVENING_21);
-    private static final Duration PERIOD = Duration.standardHours(1);
-    private static final Duration HALF_PERIOD = Duration.standardMinutes(30);
-    private static final Duration MIN_NOTIFICATION_SEPARATION = Duration.standardMinutes(30);
+    private static final Duration PERIOD = Duration.ofHours(1);
+    private static final Duration HALF_PERIOD = Duration.ofMinutes(30);
+    private static final Duration MIN_NOTIFICATION_SEPARATION = Duration.ofMinutes(30);
 
     private MockContext mContext;
     private MockTimingModelEnvironment mEnvironment;
     private TimingModel mSubject;
 
+    @SuppressWarnings("RedundantThrows")
     @Before
     public void setUp() throws Exception {
         mContext = new MockContext();
@@ -90,7 +132,7 @@ public class TimingModelTest {
         assertEquals(0, mEnvironment.sendReminderCallCount);
         assertNull(mEnvironment.setLastNotifiedTimeParams);
 
-        mEnvironment.now = MIDDAY_21.plus(Duration.standardMinutes(1));
+        mEnvironment.now = MIDDAY_21.plus(Duration.ofMinutes(1));
         mSubject.onAlarmElapsed(mContext);
         mSubject.onAlarmElapsed(mContext);
 
@@ -102,7 +144,7 @@ public class TimingModelTest {
     @Test
     public void testEnsureNotificationIsPending_nextIsSet() throws Exception {
         mEnvironment.now = MIDDAY_21;
-        mEnvironment.nextNotificationTime = MIDDAY_21.plus(Duration.standardMinutes(1));
+        mEnvironment.nextNotificationTime = MIDDAY_21.plus(Duration.ofMinutes(1));
         ensureNotificationIsPending(mSubject);
 
         assertIsEqual(mEnvironment.nextNotificationTime, mEnvironment.setNextNotificationTimeParams.alarmTime);
@@ -145,7 +187,7 @@ public class TimingModelTest {
     @Test
     public void testOnBootCompleted() {
         mEnvironment.now = MIDDAY_21;
-        mEnvironment.setLastNotifiedTime(MORNING_21.minus(Duration.standardDays(1)), mContext);
+        mEnvironment.setLastNotifiedTime(MORNING_21.minus(Duration.ofDays(1)), mContext);
         mSubject.onBootCompleted(mContext);
 
         assertEquals(1, mEnvironment.sendReminderCallCount);
@@ -155,14 +197,14 @@ public class TimingModelTest {
     @Test
     public void testOnNotified() {
         mSubject.onNotified(mContext);
-        assertIsEqual(MORNING_8AM_21, Duration.standardSeconds(1), mEnvironment.setLastNotifiedTimeParams.time);
+        assertIsEqual(MORNING_8AM_21, Duration.ofSeconds(1), mEnvironment.setLastNotifiedTimeParams.time);
         assertIsEqual(MORNING_8AM_21.plus(PERIOD), mEnvironment.setAlarmParams.alarmTime);
     }
 
     @Test
     public void testOnScreenOn() {
         mEnvironment.now = MIDDAY_21;
-        mEnvironment.setLastNotifiedTime(MORNING_21.minus(Duration.standardDays(1)), mContext);
+        mEnvironment.setLastNotifiedTime(MORNING_21.minus(Duration.ofDays(1)), mContext);
         mSubject.onScreenOn(mContext);
 
         assertEquals(1, mEnvironment.sendReminderCallCount);
@@ -189,10 +231,10 @@ public class TimingModelTest {
     }
 
     private static class MockTimingModelEnvironment implements TimingModel.Environment {
-        public DateTime now = MORNING_8AM_21;
+        public ZonedDateTime now = MORNING_8AM_21;
         boolean appInForeground = false;
-        DateTime lastNotifiedTime = null;
-        DateTime nextNotificationTime = null;
+        ZonedDateTime lastNotifiedTime = null;
+        ZonedDateTime nextNotificationTime = null;
 
         SetLastNotifiedTimeParams setLastNotifiedTimeParams = null;
         SetNextNotificationTimeParams setNextNotificationTimeParams = null;
@@ -200,12 +242,12 @@ public class TimingModelTest {
         int sendReminderCallCount = 0;
 
         @Override
-        public DateTime getNow() {
+        public ZonedDateTime getNow() {
             return now;
         }
 
         @Override
-        public DateTime getLastNotifiedTime(Context context) {
+        public ZonedDateTime getLastNotifiedTime(Context context) {
             if (lastNotifiedTime != null) {
                 return lastNotifiedTime;
             }
@@ -214,13 +256,13 @@ public class TimingModelTest {
         }
 
         @Override
-        public void setLastNotifiedTime(final DateTime time, Context context) {
+        public void setLastNotifiedTime(final ZonedDateTime time, Context context) {
             setLastNotifiedTimeParams = new SetLastNotifiedTimeParams();
             setLastNotifiedTimeParams.time = time;
         }
 
         @Override
-        public DateTime getNextNotificationTime(Context context) {
+        public ZonedDateTime getNextNotificationTime(Context context) {
             if (nextNotificationTime != null) {
                 return nextNotificationTime;
             }
@@ -229,7 +271,7 @@ public class TimingModelTest {
         }
 
         @Override
-        public void setNextNotificationTime(DateTime time, Context context) {
+        public void setNextNotificationTime(ZonedDateTime time, Context context) {
             setNextNotificationTimeParams = new SetNextNotificationTimeParams();
             setNextNotificationTimeParams.alarmTime = time;
         }
@@ -240,7 +282,7 @@ public class TimingModelTest {
         }
 
         @Override
-        public void setAlarm(DateTime alarmTime, Context context) {
+        public void setAlarm(ZonedDateTime alarmTime, Context context) {
             setAlarmParams = new SetAlarmParams();
             setAlarmParams.alarmTime = alarmTime;
         }
@@ -251,15 +293,15 @@ public class TimingModelTest {
         }
 
         public static class SetLastNotifiedTimeParams {
-            public DateTime time;
+            public ZonedDateTime time;
         }
 
         public static class SetNextNotificationTimeParams {
-            DateTime alarmTime;
+            ZonedDateTime alarmTime;
         }
 
         public static class SetAlarmParams {
-            DateTime alarmTime;
+            ZonedDateTime alarmTime;
         }
     }
 }
