@@ -1,6 +1,6 @@
-/**
-  * © Robert Williamson 2014-2016.
-  * This program is distributed under the terms of the GNU General Public License.
+/*
+   © Robert Williamson 2014-2016.
+   This program is distributed under the terms of the GNU General Public License.
   */
 package com.robwilliamson.healthyesther.unit.com.robwilliamson.healthyesther.util.time;
 
@@ -10,12 +10,12 @@ import com.robwilliamson.healthyesther.util.time.Range;
 import com.robwilliamson.healthyesther.util.time.RangeSet;
 import com.robwilliamson.healthyesther.util.time.TimeRegion;
 
-import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
-import org.joda.time.Duration;
-import org.joda.time.ReadableInstant;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import java.time.Duration;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,17 +24,43 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(AndroidJUnit4.class)
 public class RangeSetTest {
-    private static final DateTime FROM = new DateTime(2010, 4, 19, 13, 0, 0, 0, DateTimeZone.UTC);
-    private static final DateTime TO = new DateTime(2010, 4, 19, 14, 0, 0, 0, DateTimeZone.UTC);
-    private static final DateTime CENTRE = new DateTime(2010, 4, 19, 13, 30, 0, 0, DateTimeZone.UTC);
-    private static final Duration SMALL_SIGMA = Duration.standardMinutes(1);
+    private static final ZonedDateTime FROM = ZonedDateTime.of(
+            2010,
+            4,
+            19,
+            13,
+            0,
+            0,
+            0,
+            ZoneOffset.UTC);
+    private static final ZonedDateTime TO = ZonedDateTime.of(
+            2010,
+            4,
+            19,
+            14,
+            0,
+            0,
+            0,
+            ZoneOffset.UTC);
+    private static final ZonedDateTime CENTRE = ZonedDateTime.of(
+            2010,
+            4,
+            19,
+            13,
+            30,
+            0,
+            0,
+            ZoneOffset.UTC);
+    private static final Duration SMALL_SIGMA = Duration.ofMinutes(1);
 
     private static final Range SMALL_FROM_RANGE = new Range(FROM, SMALL_SIGMA);
     private static final Range SMALL_CENTRE_RANGE = new Range(CENTRE, SMALL_SIGMA);
     private static final Range SMALL_TO_RANGE = new Range(TO, SMALL_SIGMA);
     private static final Range FROM_EDGE_RANGE = new Range(FROM.minus(SMALL_SIGMA), FROM);
     private static final Range TO_EDGE_RANGE = new Range(TO, TO.plus(SMALL_SIGMA));
-    private static final Range HUGE_RANGE = new Range(FROM.minus(Duration.standardDays(1)), TO.plus(Duration.standardDays(1)));
+    private static final Range HUGE_RANGE = new Range(
+            FROM.minus(Duration.ofDays(1)),
+            TO.plus(Duration.ofDays(1)));
 
     private static final RangeSet SMALL_RANGES = new RangeSet(SMALL_FROM_RANGE, SMALL_CENTRE_RANGE, SMALL_TO_RANGE);
     private static final RangeSet SMALL_RANGES_FUTURE_EDGES = new RangeSet(
@@ -72,21 +98,26 @@ public class RangeSetTest {
     public void testStartingFromDate() {
         RangeSet subject = new RangeSet(HUGE_RANGE).startingFrom(2015, 4, 18);
         assertFalse(subject.contains(FROM));
-        assertTrue(subject.contains(FROM.withDate(2015, 4, 18)));
+        assertTrue(subject.contains(FROM
+                .withYear(2015)
+                .withMonth(4)
+                .withDayOfMonth(18)));
     }
 
     @Test
     public void testNextEdge() {
-        ReadableInstant nextEdge = SMALL_RANGES.getEdgeAfter(TO.plus(Duration.standardDays(1)));
+        ZonedDateTime nextEdge = SMALL_RANGES.getEdgeAfter(
+                TO.plus(Duration.ofDays(1)).toInstant());
         assertNull(nextEdge);
 
-        nextEdge = SMALL_RANGES.getEdgeAfter(SMALL_RANGES.to);
+        nextEdge = SMALL_RANGES.getEdgeAfter(SMALL_RANGES.to.toInstant());
         assertEquals(SMALL_RANGES.to, nextEdge);
 
-        nextEdge = SMALL_RANGES.getEdgeAfter(TO);
+        nextEdge = SMALL_RANGES.getEdgeAfter(TO.toInstant());
         assertEquals(SMALL_RANGES.to, nextEdge);
 
-        nextEdge = SMALL_RANGES.getEdgeAfter(SMALL_TO_RANGE.from.minus(Duration.standardSeconds(1)));
+        nextEdge = SMALL_RANGES.getEdgeAfter(
+                SMALL_TO_RANGE.from.minus(Duration.ofSeconds(1)).toInstant());
         assertEquals(SMALL_TO_RANGE.from, nextEdge);
     }
 }
