@@ -1,9 +1,10 @@
-/**
-  * © Robert Williamson 2014-2016.
-  * This program is distributed under the terms of the GNU General Public License.
+/*
+   © Robert Williamson 2014-2016.
+   This program is distributed under the terms of the GNU General Public License.
   */
 package com.robwilliamson.healthyesther;
 
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -11,10 +12,12 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.LinearLayout;
 
 import javax.annotation.Nullable;
 
+@SuppressLint("Registered")
 public class BaseFragmentActivity extends AppCompatActivity {
     private volatile boolean mActive = false;
 
@@ -24,7 +27,7 @@ public class BaseFragmentActivity extends AppCompatActivity {
         getWindow().setBackgroundDrawableResource(R.drawable.background);
         setContentView(getContentLayoutResourceId());
 
-        Toolbar toolbar = (Toolbar) Utils.checkNotNull(findViewById(R.id.toolbar));
+        Toolbar toolbar = (Toolbar) Utils.checkNotNull((View) findViewById(R.id.toolbar));
         setSupportActionBar(toolbar);
     }
 
