@@ -1,6 +1,6 @@
-/**
- * © Robert Williamson 2014-2016.
- * This program is distributed under the terms of the GNU General Public License.
+/*
+  © Robert Williamson 2014-2016.
+  This program is distributed under the terms of the GNU General Public License.
  */
 package com.robwilliamson.healthyesther.db.integration;
 
@@ -16,24 +16,25 @@ import com.robwilliamson.healthyesther.db.includes.DateTime;
 import com.robwilliamson.healthyesther.db.includes.WhereContains;
 import com.robwilliamson.healthyesther.util.time.Range;
 
-import org.joda.time.Duration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.MockitoAnnotations;
-import org.robolectric.RobolectricGradleTestRunner;
+import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
+import java.time.Duration;
+import java.time.ZonedDateTime;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
-@RunWith(RobolectricGradleTestRunner.class)
+@RunWith(RobolectricTestRunner.class)
 @Config(constants = BuildConfig.class)
 public class DbV4ToV5Test {
     static {
@@ -122,9 +123,9 @@ public class DbV4ToV5Test {
         HealthScoreJudgmentRangeTable.Row scoreJudgment = HealthDatabase.HEALTH_SCORE_JUDGMENT_RANGE_TABLE.select1(db, WhereContains.foreignKey(HealthScoreJudgmentRangeTable.SCORE_ID, happiness.getConcretePrimaryKey().getId()));
 
         DateTime now = DateTimeConverter.now();
-        Range range = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgment);
+        Range range = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgment);
 
-        assertThat(range.length(), equalTo(Duration.standardDays(1)));
+        assertThat(range.length(), equalTo(Duration.ofDays(1)));
     }
 
     @Test
@@ -145,15 +146,15 @@ public class DbV4ToV5Test {
         HealthScoreJudgmentRangeTable.Row[] scoreJudgments = HealthDatabase.HEALTH_SCORE_JUDGMENT_RANGE_TABLE.select(db, WhereContains.foreignKey(HealthScoreJudgmentRangeTable.SCORE_ID, drowsiness.getConcretePrimaryKey().getId()));
 
         DateTime now = DateTimeConverter.now();
-        Range day = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgments[0]);
-        Range night = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgments[1]);
+        Range day = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgments[0]);
+        Range night = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgments[1]);
 
         if (day.from.isAfter(night.to)) {
             day = night;
         }
 
-        assertThat(day.from.hourOfDay().get(), is(8));
-        assertThat(day.to.hourOfDay().get(), is(20));
+        assertThat(day.from.getHour(), is(8));
+        assertThat(day.to.getHour(), is(20));
     }
 
     @Test
@@ -164,15 +165,15 @@ public class DbV4ToV5Test {
         HealthScoreJudgmentRangeTable.Row[] scoreJudgments = HealthDatabase.HEALTH_SCORE_JUDGMENT_RANGE_TABLE.select(db, WhereContains.foreignKey(HealthScoreJudgmentRangeTable.SCORE_ID, drowsiness.getConcretePrimaryKey().getId()));
 
         DateTime now = DateTimeConverter.now();
-        Range day = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgments[0]);
-        Range night = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgments[1]);
+        Range day = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgments[0]);
+        Range night = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgments[1]);
 
         if (day.from.isAfter(night.to)) {
             night = day;
         }
 
-        assertThat(night.from.hourOfDay().get(), is(22));
-        assertThat(night.to.hourOfDay().get(), is(6));
+        assertThat(night.from.getHour(), is(22));
+        assertThat(night.to.getHour(), is(6));
     }
 
     @Test
@@ -183,8 +184,8 @@ public class DbV4ToV5Test {
         HealthScoreJudgmentRangeTable.Row scoreJudgment = HealthDatabase.HEALTH_SCORE_JUDGMENT_RANGE_TABLE.select1(db, WhereContains.foreignKey(HealthScoreJudgmentRangeTable.SCORE_ID, foonsness.getConcretePrimaryKey().getId()));
 
         DateTime now = DateTimeConverter.now();
-        Range range = Range.Starting(now.as(org.joda.time.DateTime.class)).from(scoreJudgment);
+        Range range = Range.Starting(now.as(ZonedDateTime.class)).from(scoreJudgment);
 
-        assertThat(range.length(), equalTo(Duration.standardDays(1)));
+        assertThat(range.length(), equalTo(Duration.ofDays(1)));
     }
 }
